+++ package.json
@@ -15,7 +15,7 @@
     "prettier:check": "prettier --check \"src/**/*.{js,jsx,ts,tsx,css}\"",
     "prettier:fix": "prettier --write \"src/**/*.{js,jsx,ts,tsx,css}\"",
     "eslint:fix": "eslint --ext .js,.jsx,.ts,.tsx src --fix",
-    "start": "react-scripts start",
+    "start": "node index.js",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
