+++ src/common/Button/Button.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import "./Button.css";
 
-const Button = ({ buttonText, onClick }) => {
+const Button = ({ buttonText, onClick, type = 'button' }) => {
   return (
-    <button className="button" onClick={onClick}>
+    <button className="button" onClick={onClick} type={type}>
       {buttonText}
     </button>
   );
