+++ src/common/Input/Input.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Input.css";
 
-const Input = ({ labelText, placeholderText, value, onChange }) => {
+const Input = ({ labelText, placeholderText, value, onChange, type = 'input' }) => {
   return (
     <div className="custom-input">
       <label>
@@ -11,6 +11,7 @@ const Input = ({ labelText, placeholderText, value, onChange }) => {
           placeholder={placeholderText}
           value={value}
           onChange={onChange}
+          type={type}
         />
       </label>
     </div>
